BugFix:BindableDynamicDictionary:Add type conversion to GetOrDefault
Was having a problem with model values becoming string.  Since this is all dynamic stuff, if it can be converted to the right type then convert it.  This will let us store text from textbox but treat it like an integer

The problem was happening because the counter is displayed in a textbox and was getting converted to a string then we want to pull it back out as an integer