Host Grid row height changes
If a control has a height set, which means a style was used.  All the controls have a default height of double.NaN.  The row that gets added to the grid will have the same height as the control.

Did a little work improving the look of the vertical visibility test so it would look better